bug fixes and huge simplification
I was using DataFrame.join when I should have been using DataFrame.merge.
This was causing a reporting bug and created a need for a very slow apply.
Now this is fixed and the allele_consolidator is like better than 10x faster.
And this helped me figure out a much better approach to sample_consolidator which is also probably faster